feat: Add `mesonEndpoint` config
