+++ apps/canonical-bridge-ui/pages/testnet/index.tsx
@@ -1,9 +1,42 @@
-import { CanonicalBridgeProvider, TransferWidget } from '@bnb-chain/canonical-bridge-widget';
+import {
+  CanonicalBridgeProvider,
+  ICanonicalBridgeConfig,
+  TransferWidget,
+} from '@bnb-chain/canonical-bridge-widget';
 
-import { bridgeConfig } from '@/core/config';
+import { en as messages } from '@/core/locales/en';
 import { useTestnetTransferConfig } from '@/data-testnet';
 import { testnetChains } from '@/data-testnet/testnetChains';
 import { Layout } from '@/core/components/Layout';
+import { dark } from '@/core/theme/dark';
+import { light } from '@/core/theme/light';
+import { env } from '@/core/env';
+
+export const bridgeConfig: ICanonicalBridgeConfig = {
+  appName: env.APP_NAME,
+  assetPrefix: env.ASSET_PREFIX,
+
+  appearance: {
+    bridgeTitle: 'BNB Chain Cross-Chain Bridge',
+    locale: 'en',
+    messages,
+    mode: 'dark',
+    theme: {
+      dark: dark,
+      light: light,
+    },
+  },
+  wallet: {
+    walletConnectProjectId: env.WALLET_CONNECT_PROJECT_ID,
+  },
+  http: {
+    refetchingInterval: 30 * 1000, // 30s
+    apiTimeOut: 60 * 1000, // 60s
+    deBridgeAccessToken: '',
+    serverEndpoint: env.SERVER_ENDPOINT,
+    mesonEndpoint: 'https://testnet-relayer.meson.fi/api/v1',
+  },
+};
 
 export default function TestnetPage() {
   const testnetTransferConfig = useTestnetTransferConfig();
