+++ packages/canonical-bridge-widget/src/core/hooks/useBridgeSDK.ts
@@ -14,6 +14,7 @@ export const useBridgeSDK = () => {
 
   const bridgeSDK = useMemo(() => {
     const timeout = bridgeConfig.http.apiTimeOut;
+
     return new CanonicalBridgeSDK({
       bridgeConfigs: [
         {
@@ -36,10 +37,10 @@ export const useBridgeSDK = () => {
           bridgeType: 'layerZero',
           endpoint: '',
         },
-        { bridgeType: 'meson', endpoint: 'https://testnet-relayer.meson.fi/api/v1' },
+        { bridgeType: 'meson', endpoint: bridgeConfig.http.mesonEndpoint! },
       ],
     });
-  }, [bridgeConfig.http.apiTimeOut]);
+  }, [bridgeConfig.http.apiTimeOut, bridgeConfig.http.mesonEndpoint]);
 
   return bridgeSDK;
 };
